refactor: move slash commands to new events dir
