+++ src/events/commands/createinvite.ts
@@ -1,5 +1,5 @@
 import { ChannelType, SlashCommandBuilder, type TextChannel } from 'discord.js';
-import { newSlashCommand } from '../../structures/BotClient';
+import { newSlashCommand } from '@structures/interactions/SlashCommand';
 
 const data = new SlashCommandBuilder().setName('createinvite').setDescription('See the invite links for our servers');
 data.addChannelOption((x) => x.setName('channel').setDescription('Channel to create the invite for').addChannelTypes(ChannelType.GuildText));
