+++ src/events/commands/voting/manage-votecount.ts
@@ -1,8 +1,8 @@
 import { SlashCommandBuilder } from 'discord.js';
-import { ServerType, newSlashCommand } from '../../../structures/BotClient';
 import { CustomError } from '../../../util/errors';
 import { getVoteCounter } from '@models/votecounter';
 import { genCreateVoteCountEmbed, genVoteCountEmbed } from '@views/votecounter';
+import { ServerType, newSlashCommand } from '@structures/interactions/SlashCommand';
 
 const data = new SlashCommandBuilder().setName('manage-votecount').setDescription('Commands surrounding vote counts');
