+++ tsconfig.json
@@ -22,6 +22,7 @@
 			"@utils/*": ["util/*"],
 			"@models/*": ["models/*"],
 			"@views/*": ["views/*"],
+			"@structures/*": ["structures/*"],
 			"@root/*": ["./*"]
 		}
 	},
